requirements.txt
numpy==2.2.0
psycopg2-binary==2.9.10
